requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
alembic
asyncpg
aiogram
aiosmtplib
python-dotenv
pydantic[email]
aiosqlite
